Update geocel to use constants for length units

+++ src/geocel/g4vg/SolidConverter.cc
@@ -496,7 +496,8 @@ auto SolidConverter::polyhedra(arg_type solid_base) -> result_type
         rmaxs[i] = scale_(params.Rmax[i] * radius_factor);
     }
 
-    auto phistart = std::fmod(params.Start_angle, 2 * constants::pi);
+    auto phistart = std::fmod(params.Start_angle,
+                              static_cast<double>(2 * constants::pi));
 
     return GeoManager::MakeInstance<UnplacedPolyhedron>(phistart,
                                                         params.Opening_angle,
